Securing routes

+++ server/api/remove.post.ts
@@ -1,10 +1,22 @@
 import { neon } from "@neondatabase/serverless";
+import { auth } from "~/lib/auth";
 
 export default defineEventHandler(async (event) => {
   const { DB_URL } = useRuntimeConfig(event);
   const sql = neon(DB_URL);
 
   const { type: type, id: id } = await readBody(event);
+
+  const session = await auth.api.getSession({
+    headers: event.headers,
+  });
+
+  if (!session?.user.id)
+    return createError({
+      statusCode: 401,
+      statusMessage: "Unauthorized",
+      message: "You must be logged in to perform this action.",
+    });
 
   if (type === "project") {
     await sql`DELETE FROM join_user_project WHERE project_id = ${id};`;
